Refactoring the broker to allow for install.
