--- app/manager.json
@@ -1,1 +1,0 @@
-{"URL":"http://ec2-54-210-156-182.compute-1.amazonaws.com:80"}
