+++ tests/testUtils.py
@@ -1,9 +1,9 @@
 import requests
 
-class TestUtils:
+class TestUtils(object):
     # Test basic routes, including login and file submission
-    BASE_URL = "http://127.0.0.1:80"
-    #BASE_URL = "http://54.173.199.34:80"
+    BASE_URL = "http://127.0.0.1:5000"
+    
     JSON_HEADER = {"Content-Type": "application/json"}
 
     def getRequest(self,url) :
